Introduce the multi_ops module
